Reapply "Merge branch 'main' of https://github.com/microsoft/teams-ai"
This reverts commit 42d3000a32b66fd45d9db00e1da30d04906cd0d0.

+++ python/packages/ai/teams/ai/embeddings/embeddings_response.py
@@ -24,7 +24,7 @@ class EmbeddingsResponse:
     status: EmbeddingsResponseStatus
     "Status of the embeddings response."
 
-    output: Optional[Union[List[List[int]], str]] = None
+    output: Optional[Union[List[List[float]], str]] = None
     "Optional. Embeddings for the given inputs or the error string."
 
     message: Optional[str] = None
